Changed the flow error to known-layout string
Previously, errors thrown by flow were simply numbers, with the number represening the exception's index into the exception table. This means that when flow errors are uncaught, the message shown is simply a number, along with a stack trace. To fix this, I've changed the error thrown to be a known-layout string (fixed prefix), with an easily extractable index value embedded into it. This allows more verbosity for uncaught flow exceptions.